Message: Make readVarInt() return a VarInt rather than long for Dash specific classes.

+++ core/src/main/java/org/bitcoinj/core/Headers2Message.java
@@ -107,7 +107,7 @@ public class Headers2Message extends Message {
 
     @Override
     protected void parse() throws ProtocolException {
-        long numHeaders = readVarInt();
+        long numHeaders = readVarInt().longValue();
         log.info("Parsing headers2 message: numHeaders={}, payloadLength={}", numHeaders, payload.length);
 
         if (numHeaders > MAX_HEADERS) {
